Changed book search to just search "tallaght_library" hard coded instead of looping through all libraries due to error and lack of access to a list of various libraries

+++ core/src/main/java/messages/catalogue/SearchRequest.java
@@ -16,7 +16,7 @@ public class SearchRequest {
         this.bookId = bookId;
     }
 
-    public long getBookId() {
+    public int getBookId() {
         return bookId;
     }
